refactor: extract SystemAudioTapSession from SystemAudioRecorder (#566, #435 FINAL)
Fourth and final slice of #435 (SystemAudioRecorder decomposition). Extracts
`SystemAudioTapSession` — the private final class (233 lines) that owns the
CoreAudio process-tap + aggregate-device lifecycle — into its own file.

The only intended change: visibility widen from `private` to internal (drop
`private ` modifier on class header), needed for cross-file resolution.
All method visibilities inside the class body are preserved (private helpers
stay private, no other widening).

Byte-identical body; SHA-256 33a48f98f850dcc872dda093473402c0cd172e8cbafc5d9c9d52ba0c321a31c9
verified against origin/main:SystemAudioRecorder.swift lines 262-494 (with the
`private ` prefix re-added for the hash — validates the ONLY delta is the
visibility widen).

Codex spec-review confirmed:
- No file-scope helper dependencies (uses only already-split-out symbols:
  SystemAudioAggregateDeviceCleanupSummary, SystemAudioAggregateDeviceIdentity,
  AudioObjectID extensions, SystemAudioFileWriter, AppError)
- No self-capturing `Unmanaged.passUnretained(self)` C-interop trampolines that
  would risk cross-file breakage; block callbacks capture `writer`
- @MainActor NOT propagated (SystemAudioTapSession is top-level, not @MainActor)
- DispatchQueue.drain() extension stays in SystemAudioRecorder.swift (unused by
  the tap session)

Imports: AudioToolbox + @preconcurrency AVFoundation + Foundation.

Tests: 667 (unchanged). CoreAudio tap lifecycle has no unit coverage per the
epic's baseline constraint; regression validated by full suite green.
